pyproject.toml
[tool.pysen-cli]
settings_dir = "."

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
py_version = "py38"
isort_known_first_party = ["fakes", "pysen"]
[[tool.pysen.lint.mypy_targets]]
  paths = [".", "tests/"]

[tool.pysen.lint.source]
  excludes = ["build/", "dist/", "examples/", ".tox"]

[tool.pytest.ini_options]
markers = ["examples"]

[tool.black] # automatically generated by pysen
line-length = 88
target-version = ["py38"]

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
known_first_party = ["fakes", "pysen"]
line_length = 88
multi_line_output = 3
use_parentheses = true

[tool.poetry]
name = "shinrl"
version = "0.1.0"
description = ""
authors = ["syuntoku14 <syuntoku14@gmail.com>"]
readme = "README.md"
packages = [{include = "shinrl"}]

[tool.poetry.dependencies]
python = "^3.9"
gym = "0.23.1"
jax = "0.4.1"
jaxlib = "0.4.1"
tqdm = "^4.65.0"
cpprb = "10.6.4"
structlog = "^23.1.0"
colorama = "^0.4.6"
pyyaml = "^6.0"
seaborn = "0.12.1"
matplotlib = "3.6.2"
chex = "0.1.5"
dm-haiku = "0.0.9"
optax = "0.1.4"
distrax = "0.1.2"
celluloid = "^0.2.0"
pytest = "^7.3.1"
pysen = {extras = ["lint"], version = "^0.10.4"}
pandas = "1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
